Declare the interfaces the geo and http fakes stand in for
Of the 31 test errors left after the page protocol, nine were fakes with nowhere to go: FakeGeo passed as a GeoService, ScriptedGeocoder as a geopy Nominatim, FakeHttpClient as an httpx client.

geo.py now declares what it asks a geocoder for — LocationLike, GeocoderLike, GeoServiceLike — and discovery.py declares the two members it uses of a client, ResponseLike and ClientLike. The geo ones are not decoration: geopy *generates* geocode behind a decorator that switches on the adapter (base.py has the synchronous def commented out), so a synchronous call read as a coroutine and location.latitude as unknown. The four geo.py artifacts §8.9 used to record as unfixable are gone, and the call site is checked instead of excused; one cast, where the Nominatim is built, is what bridges the decorator (§15.13). GeoServiceLike is deliberately not a GeocoderLike — the facade returns tuples, not Locations — and my first attempt that inherited one from the other was rightly rejected.

The rest were the tests' own, each fixed where it went wrong: a MagicMock read in a different method than the one that installed it (MSAL's app, where the narrowing does not cross methods — one _msal_app helper with one cast; and monitor.delete, now held by name as delete_mock); a SimpleNamespace standing in for the app, cast at the single call; a handler dict declared dict[str, object] and then called; a FakeContext of its own in test_registrar_setup, which had grown four duplicates of the shared fakes — deleted, since tests/fake_async_page.py exists to stop exactly that; and a faked playwright, cast in the one helper that installs it.

The scheduler got no protocol on purpose. Its stub stands in for OilWatchApp, our own class, and an eight-member protocol would be a hand-maintained copy of our own API that would stop checking the scheduler against the app it actually drives — one cast in one _wired helper is the smaller trade (§15.16).

Coverage measures 95.52, and 94.75 with the protocol bodies counted: a stub body can never be entered, so its entry branch counts as partial for good, and the drag grows with every protocol added. exclude_lines now says that an ellipsis is a declaration, which is true of every one of them in this package. Omitting the files instead would have hidden real statements if they ever gain any (§15.17).

pyright tests: 48 errors to 0. Bare pyright: nine to five (analytics x4, import_xls x1). docs/inspection.md records the new baseline in 1.10, 8.9, 10.2 and 10.4, with 15.15 to 15.17 for what this taught.

623 tests pass; ruff clean.

+++ tests/test_geo.py
@@ -21,7 +21,7 @@ class ScriptedGeocoder:
         self._results = results
         self.queries: list[str] = []
 
-    def geocode(self, query: str, timeout: int | None = None):
+    def geocode(self, query: str, timeout: float | None = None):
         self.queries.append(query)
         result = self._results.get(query)
         if isinstance(result, Exception):
